release: Build smaller examples
These are low-hanging fruits towards having compact binaries with much
inlining. This change is not so much for performance or binary size
itself (I haven't done benchmarks there), but to see in the disassembled
binary where abstractions are actually zero-cost. (Eg. "Is setting the
GPIO pin actually just the required instruction?").